requirements.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
cryptography==2.9.2
Django==2.2.10
django-cors-headers==3.2.1
django-daterange-filter==1.3.0
django-jet==1.0.8
django-qiniu-storage==2.3.1
django-redis==4.11.0
idna==2.8
jwt==0.6.1
mysqlclient==1.4.6
Pillow==7.0.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
qiniu==7.2.4
redis==3.4.1
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
xmltodict==0.12.0
